refactor: 유틸함수들을 index.js에 묶어서 export한다
