+++ src/utils/index.js
@@ -1,0 +1,8 @@
+import { LOTTO, MESSAGE } from './constants';
+import deepFreeze from './deepFreeze';
+import formValidator from './formValidator';
+import inputValidator from './inputValidator';
+import generateId from './generateId';
+import generateLottoNumbers from './generateTicket';
+
+export { LOTTO, MESSAGE, deepFreeze, formValidator, inputValidator, generateId, generateLottoNumbers };
